requirements.txt
absl-py==2.3.1
accelerate==1.6.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.10.0
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.1.0
audioread==3.0.1
av==14.1.0
beautifulsoup4==4.13.4
blessed==1.20.0
blobfile==3.0.0
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
cfgv==3.4.0
click==8.1.8
cloudpickle==3.1.1
cn2an==0.5.23
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
conformer==0.3.2
contourpy==1.3.1
contractions==0.1.73
crcmod==1.7
cryptography==44.0.0
cssselect2==0.8.0
ctranslate2==4.4.0
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.14
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.7
defusedxml==0.7.1
diffusers==0.33.1
dill==0.3.8
distlib==0.4.0
docopt==0.6.2
easydict==1.13
editdistance==0.8.1
einops==0.8.0
evaluate==0.4.3
executing==2.2.0
fastapi==0.116.1
fastcluster==1.2.6
faster-whisper==1.1.0
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.17.0
fire==0.7.0
flatbuffers==25.1.24
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.9.0
funasr==1.2.7
funasr-onnx==0.4.1
future==1.0.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gpustat==1.1.1
gradio==5.45.0
gradio_client==1.13.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.7
hjson==3.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
identify==2.6.13
imagesize==1.4.1
immutabledict==4.2.1
importlib_metadata==8.7.0
indic_numtowords==1.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.10.0
jiwer==3.1.0
jmespath==0.10.0
joblib==1.4.2
json_repair==0.48.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kaldi-native-fbank==1.20.2
kaldiio==2.18.0
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.22.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
onnxruntime-gpu==1.20.1
openai==1.99.5
openai-whisper==20250625
opencv-python-headless==4.11.0.86
optuna==4.2.0
orjson==3.11.3
oss2==2.19.1
pandas==2.2.3
panns-inference==0.1.1
parso==0.8.4
pastel==0.2.1
pexpect==4.9.0
pillow==11.1.0
poethepoet==0.10.0
pooch==1.8.2
posix_ipc==1.3.0
pre_commit==4.3.0
primePy==1.3
proces==0.1.7
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.26.1
protobuf>=3.20.2,<6.0.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyahocorasick==2.2.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.21.0
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
qa_metrics==0.2.42
qwen-omni-utils==0.0.8
RapidFuzz==3.13.0
ray==2.41.0
referencing==0.36.2
regex==2024.11.6
rich==13.9.4
rotary-embedding-torch==0.8.9
rpds-py==0.22.3
rsa==4.9.1
ruff==0.13.0
s3prl==0.4.18
s3tokenizer==0.2.0
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.15.1
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.47.3
stopes==2.2.1
submitit==1.5.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.2.2
termcolor==3.1.0
textsearch==0.0.24
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.4
tomlkit==0.13.2
torch==2.6.0
torch-audiomentations==0.12.0
torch-complex==0.4.4
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchcde==0.2.5
torchdiffeq==0.2.5
torchdyn==1.0.6
torchlibrosa==0.1.0
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.21.0
tornado==6.5.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.49.0
triton==3.2.0
typer==0.15.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
umap-learn==0.5.7
uritemplate==4.2.0
uvicorn==0.35.0
virtualenv==20.34.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wget==3.2
wheel==0.45.1
whisper_normalizer==0.1.12
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.18.3
zhconv==1.4.3
zhon==2.1.1
zipp==3.21.0
zstandard==0.23.0
